Port the coherence sweep, IRLS fit and seeding; seeds match to 1e-10 s
The second half of Phase 1's core. overtone-tempo takes attack times and
weights and returns fitted grids -- no file handling, no audio decoding,
nothing that could make a change here depend on a change elsewhere. It is
the code that produces 0.16 ms offsets.

Validated against v3 on all 24 fixtures: the anchor seed's period lands
within 1e-11 to 2.6e-7 s, against a 1e-6 s tolerance (1e-6 s at 340 ms
per beat is about 0.0002 BPM). Attacks still exact.

The delicate parts are ported with their reasoning attached, because each
is easy to lose and none announces itself when lost:

  - the coherence phase sign, phi = arg(z)/(2*pi*f). v3's timeline calls
    the negated version its single largest error source: every seed grid
    started in anti-phase and least squares cannot recover a slipped beat
    index. It gets the first test in the file.
  - recentre_phase at pass index 2 exactly, taking the MODE rather than
    the mean, which is what stops swung off-beats dragging the offset.
  - share x (0.25 + 0.75 * coverage) ranking, with the two alternatives
    v3 tried and rejected recorded so they are not re-litigated.
  - the [0.5x, 2x] guard that stops a pass octave-hopping mid-refinement.

The candidate list turned out NOT to be a valid invariant, and finding
that out is the useful part. Two fixtures had zero candidates matching
v3's to 1e-6 s while their seeds matched to 8e-8. The sweep resolves
frequency to 0.2/span, and where two adjacent grid points score almost
equally, a difference in summation order -- NumPy dispatches the dot
product to BLAS, this crate sums in attack order -- moves the peak one
index. That shifts a candidate period by up to 4e-5 s and its phase by
9e-4 s, since phase is arg(z)/(2*pi*f) and a 0.001 Hz shift accumulates
over a 29 s window. Gating on peak indices was measuring something the
stage cannot resolve, so the gate is now the property that does have to
hold: the candidate list must CONTAIN the grid v3 went on to seed.

Two of my own tests were also wrong and are fixed rather than worked
around: they expected `refine` alone to absorb a 1 % seed error over 200
beats, which slips indices by more than a period. A real coherence seed
is ~0.1 % off; the 1 % case is what `expand` exists for, and it has its
own test.

peaks is now generic over f32 and f64 -- the envelope is f32 (numpy's
float32, and the type matters for ties) and the coherence curve is f64 --
so the scipy semantics are not duplicated and cannot drift apart.

Also added: a `candidates <case>` debug command that prints both lists
side by side, which is how the above was diagnosed.

64 Rust tests. Corpus still 2.7 s against Python's 21.6 s.

+++ crates/overtone-dsp/src/lib.rs
@@ -35,7 +35,7 @@ pub fn detect_attacks(y: &[f32], sr: u32, hop: usize, retime_attacks: bool) -> (
     // bare-distance retry when nothing survives.
     let distance = ((0.025 * sr as f64 / hop as f64).round() as usize).max(1);
     let env64: Vec<f64> = env.iter().map(|&v| v as f64).collect();
-    let floor = 0.04f32.max(envelope::percentile(&env64, 55.0) as f32);
+    let floor = 0.04f64.max(envelope::percentile(&env64, 55.0));
     let mut picked = peaks::find_peaks(&env, distance, Some(0.05), Some(floor));
     if picked.is_empty() {
         picked = peaks::find_peaks(&env, distance, None, None);
